+++ android/engine/src/test/java/org/smartregister/fhircore/engine/task/FhirCarePlanGeneratorTest.kt
@@ -637,7 +637,8 @@ class FhirCarePlanGeneratorTest : RobolectricTest() {
               this.all { it.basedOn.first().reference == carePlan.asReference().reference }
             )
 
-            // first visit is lmp plus 1 month and subsequent visit are every month after that until
+            // first visit is lmp plus 1 month and subsequent visit are every month after
+            // that until
             // delivery
             val pregnancyStart = lmp.clone() as Date
             this.forEachIndexed { index, task ->
@@ -702,7 +703,8 @@ class FhirCarePlanGeneratorTest : RobolectricTest() {
               this.all { it.basedOn.first().reference == carePlan.asReference().reference }
             )
 
-            // first visit is lmp plus 1 month and subsequent visit are every month after that until
+            // first visit is lmp plus 1 month and subsequent visit are every month after
+            // that until
             // delivery
             // skip tasks for past 3 months of late registration
             val ancStart = lmp.value.plusMonths(3).clone() as Date
@@ -864,6 +866,118 @@ class FhirCarePlanGeneratorTest : RobolectricTest() {
 
               val task = tasks.find { it.description.startsWith(vaccine.key) }
               assertNotNull(task)
+              assertTrue(task!!.executionPeriod.start.asYyyyMmDd() == vaccine.value.asYyyyMmDd())
+            }
+          }
+      }
+  }
+
+  @Test
+  @ExperimentalCoroutinesApi
+  fun `Generate CarePlan should generate child immunization schedule with correct groupIdentifier value`() =
+      runTest {
+    val planDefinitionResources =
+      loadPlanDefinitionResources("child-immunization-schedule", listOf("register-temp"))
+    val planDefinition = planDefinitionResources.planDefinition
+    val patient = planDefinitionResources.patient
+    val questionnaireResponses = planDefinitionResources.questionnaireResponses
+    val resourcesSlot = planDefinitionResources.resourcesSlot
+
+    fhirCarePlanGenerator.generateOrUpdateCarePlan(
+        planDefinition,
+        patient,
+        Bundle().addEntry(Bundle.BundleEntryComponent().apply { resource = patient })
+      )!!
+      .also { println(it.encodeResourceToString()) }
+      .also { carePlan ->
+        assertCarePlan(
+          carePlan,
+          planDefinition,
+          patient,
+          patient.birthDate,
+          patient.birthDate.plusDays(4017),
+          20
+        )
+
+        resourcesSlot
+          .filter { res -> res.resourceType == ResourceType.Task }
+          .map {
+            println(it.encodeResourceToString())
+            it as Task
+          }
+          .also { tasks ->
+            assertTrue(tasks.all { it.status == TaskStatus.REQUESTED })
+            assertTrue(
+              tasks.all {
+                it.reasonReference.reference == "Questionnaire/9b1aa23b-577c-4fb2-84e3-591e6facaf82"
+              }
+            )
+            assertTrue(
+              tasks.all {
+                it.code.codingFirstRep.display ==
+                  "Administration of vaccine to produce active immunity (procedure)" &&
+                  it.code.codingFirstRep.code == "33879002"
+              }
+            )
+            assertTrue(tasks.all { it.description.contains(it.reasonCode.text, true) })
+            assertTrue(
+              tasks.all { it.`for`.reference == questionnaireResponses.first().subject.reference }
+            )
+            assertTrue(
+              tasks.all { it.basedOnFirstRep.reference == carePlan.asReference().reference }
+            )
+          }
+          .also { tasks ->
+            val vaccines =
+              mutableMapOf<String, Date>(
+                "BCG" to patient.birthDate,
+                "OPV 0" to patient.birthDate,
+                "PENTA 1" to patient.birthDate.plusDays(42),
+                "OPV 1" to patient.birthDate.plusDays(42),
+                "PCV 1" to patient.birthDate.plusDays(42),
+                "ROTA 1" to patient.birthDate.plusDays(42),
+                "PENTA 2" to patient.birthDate.plusDays(70),
+                "OPV 2" to patient.birthDate.plusDays(70),
+                "PCV 2" to patient.birthDate.plusDays(70),
+                "ROTA 2" to patient.birthDate.plusDays(70),
+                "PENTA 3" to patient.birthDate.plusDays(98),
+                "OPV 3" to patient.birthDate.plusDays(98),
+                "PCV 3" to patient.birthDate.plusDays(98),
+                "IPV" to patient.birthDate.plusDays(98),
+                "MEASLES 1" to patient.birthDate.plusMonths(9),
+                "MEASLES 2" to patient.birthDate.plusMonths(15),
+                "YELLOW FEVER" to patient.birthDate.plusMonths(9),
+                "TYPHOID" to patient.birthDate.plusMonths(9),
+                "HPV 1" to patient.birthDate.plusDays(3285),
+                "HPV 2" to patient.birthDate.plusDays(3467),
+              )
+            vaccines.forEach { vaccine ->
+              println(vaccine)
+
+              val task = tasks.find { it.description.startsWith(vaccine.key) }
+              assertNotNull(task)
+              when (vaccine.key) {
+                "BCG" -> assertTrue(task!!.groupIdentifier.value == "0_d")
+                "OPV 0" -> assertTrue(task!!.groupIdentifier.value == "0_d")
+                "PENTA 1" -> assertTrue(task!!.groupIdentifier.value == "6_wk")
+                "OPV 1" -> assertTrue(task!!.groupIdentifier.value == "6_wk")
+                "PCV 1" -> assertTrue(task!!.groupIdentifier.value == "6_wk")
+                "ROTA 1" -> assertTrue(task!!.groupIdentifier.value == "6_wk")
+                "PENTA 2" -> assertTrue(task!!.groupIdentifier.value == "10_wk")
+                "OPV 2" -> assertTrue(task!!.groupIdentifier.value == "10_wk")
+                "PCV 2" -> assertTrue(task!!.groupIdentifier.value == "10_wk")
+                "ROTA 2" -> assertTrue(task!!.groupIdentifier.value == "10_wk")
+                "PENTA 3" -> assertTrue(task!!.groupIdentifier.value == "14_wk")
+                "OPV 3" -> assertTrue(task!!.groupIdentifier.value == "14_wk")
+                "PCV 3" -> assertTrue(task!!.groupIdentifier.value == "14_wk")
+                "IPV" -> assertTrue(task!!.groupIdentifier.value == "14_wk")
+                "MEASLES 1" -> assertTrue(task!!.groupIdentifier.value == "9_mo")
+                "MEASLES 2" -> assertTrue(task!!.groupIdentifier.value == "15_mo")
+                "YELLOW FEVER" -> assertTrue(task!!.groupIdentifier.value == "9_mo")
+                "TYPHOID" -> assertTrue(task!!.groupIdentifier.value == "9_mo")
+                "HPV 1" -> assertTrue(task!!.groupIdentifier.value == "3285_d")
+                "HPV 2" -> assertTrue(task!!.groupIdentifier.value == "3467_d")
+              }
               assertTrue(task!!.executionPeriod.start.asYyyyMmDd() == vaccine.value.asYyyyMmDd())
             }
           }